[aquarium]: Improve fish path algorithm
Fish now do 3 distinct steps during their update method:

- (when possible) Try to follow their path
- (when possible) Try to target some food
- (20%) Path to the closest fish
- (30%) Idle (maybe look around)
- (50%) Path to the aquarium's current POI

The aquarium's POI is also new. It is a random point within bounds,
that gets updated when the number of calls to the `get_poi` function
are larger than the result of its `poi_update_percent` multiplied by
the number of fish it has.